+++ sources/plugins/Linky/Linky.cpp
@@ -3,8 +3,8 @@
 #include <shared/event/EventTimer.h>
 #include <plugin_cpp_api/ImplementationHelper.h>
 #include <shared/communication/PortException.hpp>
-#include "LinkyFactory.h"
 #include <shared/Log.h>
+#include "GPIOManager.hpp"
 
 // Shortcut to yadomsApi namespace
 namespace yApi = shared::plugin::yPluginApi;
@@ -16,14 +16,13 @@ namespace yApi = shared::plugin::yPluginApi;
 IMPLEMENT_PLUGIN(CLinky)
 
 CLinky::CLinky():
+   m_configuration(boost::make_shared<CLinkyConfiguration>()),
    m_isDeveloperMode(false),
    m_runningState(ELinkyPluginState::kUndefined)
-{
-}
+{}
 
 CLinky::~CLinky()
-{
-}
+{}
 
 // Event IDs
 enum
@@ -38,35 +37,38 @@ enum
 void CLinky::doWork(boost::shared_ptr<yApi::IYPluginApi> api)
 {
    YADOMS_LOG(information) << "Linky is starting..." ;
-
    m_isDeveloperMode = api->getYadomsInformation()->developperMode();
 
    // Load configuration values (provided by database)
-   m_configuration.initializeWith(api->getConfiguration());
+   m_configuration->initializeWith(api->getConfiguration());
 
-   // Create the transceiver
-   m_decoder = CLinkyFactory::constructDecoder(api);
+   m_GPIOManager = boost::make_shared<CGPIOManager>(m_configuration);
 
-   // Create the buffer handler
-   m_receiveBufferHandler = CLinkyFactory::GetBufferHandler(api->getEventHandler(),
-                                                            kEvtPortDataReceived,
-                                                            m_isDeveloperMode);
+   // For port 0 & 1
+   m_protocolManager[0] = boost::make_shared<CProtocolManager>();
+   m_protocolManager[1] = boost::make_shared<CProtocolManager>();
+
+   // Create decoders
+   m_decoder.insert(std::pair<EProtocolType, boost::shared_ptr<IDecoder>>(EProtocolType::Historic, CLinkyFactory::constructDecoder(EProtocolType::Historic, api)));
+   m_decoder.insert(std::pair<EProtocolType, boost::shared_ptr<IDecoder>>(EProtocolType::Standard, CLinkyFactory::constructDecoder(EProtocolType::Standard, api)));
 
    m_waitForAnswerTimer = api->getEventHandler().createTimer(kAnswerTimeout,
                                                              shared::event::CEventTimer::kOneShot,
-                                                             boost::posix_time::seconds(20));
+                                                             boost::posix_time::seconds(5));
 
    m_periodicSamplingTimer = api->getEventHandler().createTimer(kSamplingTimer,
                                                                 shared::event::CEventTimer::kPeriodic,
                                                                 boost::posix_time::seconds(30));
 
+   m_waitForAnswerTimer->stop();
+
+   // Create the connection
+   createConnection(api);
+
    // For immediat sampling
    m_periodicSamplingTimer = api->getEventHandler().createTimer(kSamplingTimer,
                                                                 shared::event::CEventTimer::kOneShot,
                                                                 boost::posix_time::seconds(0));
-
-   // Create the connection
-   createConnection(api);
 
    // the main loop
    YADOMS_LOG(information) << "Linky plugin is running..." ;
@@ -105,26 +107,42 @@ void CLinky::doWork(boost::shared_ptr<yApi::IYPluginApi> api)
       case kEvtPortDataReceived:
          {
 			 m_waitForAnswerTimer->stop();
-
-            if (m_isDeveloperMode) YADOMS_LOG(information) << "Linky plugin :  DataReceived" ;
+          m_protocolManager[m_GPIOManager->getGPIO()-1]->validateProtocol();
+          YADOMS_LOG(trace) << "Linky plugin :  DataReceived";
 
             processDataReceived(api,
                                 api->getEventHandler().getEventData<boost::shared_ptr<std::map<std::string, std::vector<std::string>>>>());
 
+            CLinkyFactory::FTDI_DisableGPIO(m_port);
             m_receiveBufferHandler->desactivate();
+
+            if (!m_GPIOManager->isLast())
+            {
+               // Go to the next GPIO & fire it rapidly
+               m_GPIOManager->next();
+               api->getEventHandler().createTimer(kSamplingTimer,
+                                                  shared::event::CEventTimer::kOneShot,
+                                                  boost::posix_time::seconds(1));
+            }
+            else // Return to the beginning of the list
+               m_GPIOManager->front();
+
             break;
          }
       case yApi::IYPluginApi::kEventUpdateConfiguration:
          {
-            m_runningState = kupdateConfiguration;
-            api->setPluginState(yApi::historization::EPluginState::kCustom, "updateConfiguration");
+            setPluginState(api, kupdateConfiguration);
             onUpdateConfiguration(api, api->getEventHandler().getEventData<shared::CDataContainer>());
-            api->setPluginState(yApi::historization::EPluginState::kRunning);
-
+            setPluginState(api, kRunning);
             break;
          }
       case kSamplingTimer:
       {
+         auto detectedProtocol = m_protocolManager[m_GPIOManager->getGPIO() - 1]->getProtocol();
+         CLinkyFactory::FTDI_setNewProtocol(m_port, detectedProtocol);
+         CLinkyFactory::FTDI_ActivateGPIO(m_port, m_GPIOManager->getGPIO());
+
+         m_receiveBufferHandler->changeProtocol(detectedProtocol);
          m_receiveBufferHandler->activate();
 
          //Lauch a new time the time out to detect connexion failure
@@ -133,14 +151,39 @@ void CLinky::doWork(boost::shared_ptr<yApi::IYPluginApi> api)
       }
       case kErrorRetryTimer:
          {
-            createConnection(api);
+            try {
+               m_protocolManager[m_GPIOManager->getGPIO() - 1]->changeProtocol();
+               createConnection(api);
+            }
+            catch (...)
+            {
+               YADOMS_LOG(trace) << "create connexion error";
+               errorProcess(api);
+            }
             break;
          }
       case kAnswerTimeout:
          {
-            m_waitForAnswerTimer->stop();
-            YADOMS_LOG(error) << "No answer received, try to reconnect in a while..." ;
-            errorProcess(api);
+            m_periodicSamplingTimer->stop();
+            CLinkyFactory::FTDI_DisableGPIO(m_port);
+            YADOMS_LOG(information) << "No answer received, try to reconnect in a while..." ;
+
+            if (!m_protocolManager[m_GPIOManager->getGPIO() - 1]->end())
+               changeProtocol(api);
+            else
+            {
+               errorProcess(api);
+
+               //Reset the protocol manager of the selected GPIO
+               if (!m_protocolManager[m_GPIOManager->getGPIO() - 1]->isValidated())
+                  m_protocolManager[m_GPIOManager->getGPIO() - 1]->changeProtocol();
+
+               // We continue to scan the other port if any
+               if (!m_GPIOManager->isLast())
+                  m_GPIOManager->next();
+               else
+                  m_GPIOManager->front();
+            }
             break;
          }
       default:
@@ -154,21 +197,30 @@ void CLinky::doWork(boost::shared_ptr<yApi::IYPluginApi> api)
 
 void CLinky::createConnection(boost::shared_ptr<yApi::IYPluginApi> api)
 {
-   api->setPluginState(yApi::historization::EPluginState::kCustom, "connecting");
+   setPluginState(api, kConnecting);
+
+   auto detectedProtocol = m_protocolManager[m_GPIOManager->getGPIO() - 1]->getProtocol();
+   m_receiveBufferHandler = CLinkyFactory::GetBufferHandler(detectedProtocol,
+                                                            api->getEventHandler(),
+                                                            kEvtPortDataReceived,
+                                                            m_isDeveloperMode);
+
    // Create the port instance
-   m_port = CLinkyFactory::constructPort(m_configuration,
+   m_port = CLinkyFactory::constructPort(detectedProtocol,
+                                         m_configuration,
                                          api->getEventHandler(),
                                          m_receiveBufferHandler,
                                          kEvtPortConnection);
+
    m_port->start();
-   m_waitForAnswerTimer->start();
+   m_periodicSamplingTimer->start();
 }
 
 void CLinky::destroyConnection()
 {
-   m_port.reset();
-
+   m_periodicSamplingTimer->stop();
    m_waitForAnswerTimer->stop();
+   m_port.reset();
 }
 
 void CLinky::onUpdateConfiguration(boost::shared_ptr<yApi::IYPluginApi> api,
@@ -176,6 +228,7 @@ void CLinky::onUpdateConfiguration(boost::shared_ptr<yApi::IYPluginApi> api,
 {
    // Stop running timers, if any
    m_waitForAnswerTimer->stop();
+   m_periodicSamplingTimer->stop();
 
    // Configuration was updated
    YADOMS_LOG(information) << "Update configuration..." ;
@@ -185,30 +238,38 @@ void CLinky::onUpdateConfiguration(boost::shared_ptr<yApi::IYPluginApi> api,
    if (!m_port)
    {
       // Update configuration
-      m_configuration.initializeWith(newConfigurationData);
+      m_configuration->initializeWith(newConfigurationData);
+      m_GPIOManager = boost::make_shared<CGPIOManager>(m_configuration);
       return;
    }
+   else
+   {
+      // Port has changed, destroy and recreate connection
+      destroyConnection();
 
-   // Port has changed, destroy and recreate connection
-   destroyConnection();
+      // Update configuration
+      m_configuration->initializeWith(newConfigurationData);
+      m_GPIOManager = boost::make_shared<CGPIOManager>(m_configuration);
 
-   // Update configuration
-   m_configuration.initializeWith(newConfigurationData);
+      // Create a new time protocol manager, to reset them
+      m_protocolManager[0] = boost::make_shared<CProtocolManager>();
+      m_protocolManager[1] = boost::make_shared<CProtocolManager>();
 
-   // Create new connection
-   createConnection(api);
+      // Create new connection
+      createConnection(api);
+   }
 }
 
 void CLinky::processDataReceived(boost::shared_ptr<yApi::IYPluginApi> api,
                                  const boost::shared_ptr<std::map<std::string, std::vector<std::string>>>& messages)
 {
-   m_decoder->decodeLinkyMessage(api, messages);
+   auto detectedProtocol = m_protocolManager[m_GPIOManager->getGPIO() - 1]->getProtocol();
 
-   if (m_runningState != kRunning)
-   {
-      api->setPluginState(yApi::historization::EPluginState::kRunning);
-      m_runningState = kRunning;
-   }
+   // Select the decoder from the protocol detected
+   m_decoder[detectedProtocol]->decodeMessage(api, messages);
+
+   if (!m_decoder[detectedProtocol]->isERDFCounterDesactivated())
+      setPluginState(api, kRunning);
 }
 
 void CLinky::processLinkyConnectionEvent(boost::shared_ptr<yApi::IYPluginApi> api) const
@@ -230,21 +291,26 @@ void CLinky::processLinkyConnectionEvent(boost::shared_ptr<yApi::IYPluginApi> ap
 void CLinky::processLinkyUnConnectionEvent(boost::shared_ptr<yApi::IYPluginApi> api, boost::shared_ptr<shared::communication::CAsyncPortConnectionNotification> notification)
 {
    YADOMS_LOG(information) << "Linky connection was lost" ;
-   if(notification)
+   if (notification)
       api->setPluginState(yApi::historization::EPluginState::kError, notification->getErrorMessageI18n(), notification->getErrorMessageI18nParameters());
    else
-      api->setPluginState(yApi::historization::EPluginState::kError, "connectionLost");
+      setPluginState(api, kConnectionLost);
    m_runningState = kConnectionLost;
 
    destroyConnection();
 }
 
+void CLinky::changeProtocol(boost::shared_ptr<yApi::IYPluginApi> api)
+{
+   destroyConnection();
+   api->getEventHandler().createTimer(kErrorRetryTimer, shared::event::CEventTimer::kOneShot, boost::posix_time::seconds(5));
+}
+
 void CLinky::errorProcess(boost::shared_ptr<yApi::IYPluginApi> api)
 {
-   api->setPluginState(yApi::historization::EPluginState::kError, "connectionLost");
    destroyConnection();
-   m_runningState = kConnectionLost;
-   api->getEventHandler().createTimer(kErrorRetryTimer, shared::event::CEventTimer::kOneShot, boost::posix_time::seconds(30));
+   setPluginState(api, kConnectionLost);
+   api->getEventHandler().createTimer(kErrorRetryTimer, shared::event::CEventTimer::kOneShot, boost::posix_time::seconds(5));
 }
 
 void CLinky::initLinkyReceiver() const
@@ -253,4 +319,48 @@ void CLinky::initLinkyReceiver() const
 
    // Flush receive buffer
    m_port->flush();
+}
+
+void CLinky::setPluginState(boost::shared_ptr<yApi::IYPluginApi> api, ELinkyPluginState newState)
+{
+   if (m_runningState != newState)
+   {
+      switch (newState)
+      {
+      case ELinkyPluginState::kErDFCounterdesactivated:
+         api->setPluginState(yApi::historization::EPluginState::kCustom, "ErDFCounterdesactivated");
+         break;
+      case ELinkyPluginState::kupdateConfiguration:
+         api->setPluginState(yApi::historization::EPluginState::kCustom, "updateconfiguration");
+         break;
+      case ELinkyPluginState::kConnectionLost:
+         api->setPluginState(yApi::historization::EPluginState::kError, "connectionLost");
+         break;
+      case ELinkyPluginState::kConnecting:
+         api->setPluginState(yApi::historization::EPluginState::kCustom, "connecting");
+         break;
+      case ELinkyPluginState::kRunning:
+         api->setPluginState(yApi::historization::EPluginState::kRunning);
+         break;
+      case ELinkyPluginState::kStop:
+         api->setPluginState(yApi::historization::EPluginState::kStopped);
+         break;
+      default:
+         YADOMS_LOG(error) << "this plugin status does not exist : " << newState;
+         break;
+      }
+
+      m_runningState = newState;
+   }
+}
+
+void CLinky::setPluginErrorState(boost::shared_ptr<yApi::IYPluginApi> api,
+                                 const std::string& ErrorMessageI18n,
+                                 const std::map<std::string, std::string>& ErrorMessageI18nParameters)
+{
+   if (m_runningState != kError)
+   {
+      m_runningState = kError;
+      api->setPluginState(yApi::historization::EPluginState::kError, ErrorMessageI18n, ErrorMessageI18nParameters);
+   }
 }